refactor(media): extract Gallery + supporting components into shared module
